Move shared RequestOptions override code into single class.
