+++ annotation/compiler/src/main/java/com/bumptech/glide/annotation/compiler/RequestBuilderGenerator.java
@@ -2,13 +2,11 @@ package com.bumptech.glide.annotation.compiler;
 
 import static com.bumptech.glide.annotation.compiler.ProcessorUtil.checkResult;
 import static com.bumptech.glide.annotation.compiler.ProcessorUtil.nonNull;
-import static com.bumptech.glide.annotation.compiler.RequestOptionsGenerator.BASE_REQUEST_OPTIONS_QUALIFIED_NAME;
 
 import com.bumptech.glide.annotation.GlideExtension;
 import com.bumptech.glide.annotation.GlideOption;
 import com.google.common.base.Function;
 import com.google.common.base.Joiner;
-import com.google.common.base.Predicate;
 import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
@@ -24,12 +22,7 @@ import com.squareup.javapoet.TypeSpec;
 import com.squareup.javapoet.TypeVariableName;
 import com.squareup.javapoet.WildcardTypeName;
 import java.io.File;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-import javax.annotation.Nullable;
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.ExecutableElement;
@@ -114,13 +107,11 @@ final class RequestBuilderGenerator {
 
   private final ProcessingEnvironment processingEnv;
   private final ProcessorUtil processorUtil;
-  private final TypeElement baseRequestOptionsType;
-  private ClassName generatedRequestBuilderClassName;
+  private final RequestOptionsOverrideGenerator requestOptionsOverrideGenerator;
   private final TypeVariableName transcodeTypeName;
-  private ParameterizedTypeName generatedRequestBuilderOfTranscodeType;
   private final TypeElement requestOptionsType;
   private final TypeElement requestBuilderType;
-  private ClassName requestOptionsClassName;
+  private ClassName generatedRequestBuilderClassName;
 
   RequestBuilderGenerator(ProcessingEnvironment processingEnv, ProcessorUtil processorUtil) {
     this.processingEnv = processingEnv;
@@ -133,25 +124,16 @@ final class RequestBuilderGenerator {
 
     requestOptionsType = processingEnv.getElementUtils().getTypeElement(
         REQUEST_OPTIONS_QUALIFIED_NAME);
-    baseRequestOptionsType = processingEnv.getElementUtils().getTypeElement(
-        BASE_REQUEST_OPTIONS_QUALIFIED_NAME);
+
+    requestOptionsOverrideGenerator =
+        new RequestOptionsOverrideGenerator(processingEnv, processorUtil);
   }
 
-  TypeSpec generate(String generatedCodePackageName, @Nullable TypeSpec generatedOptions) {
+  TypeSpec generate(String generatedCodePackageName) {
     generatedRequestBuilderClassName =
         ClassName.get(generatedCodePackageName, GENERATED_REQUEST_BUILDER_SIMPLE_NAME);
-    generatedRequestBuilderOfTranscodeType =
+    ParameterizedTypeName generatedRequestBuilderOfTranscodeType =
         ParameterizedTypeName.get(generatedRequestBuilderClassName, transcodeTypeName);
-
-    if (generatedOptions != null) {
-      requestOptionsClassName =
-          ClassName.get(generatedCodePackageName, generatedOptions.name);
-    } else {
-      requestOptionsClassName =
-          ClassName.get(
-              RequestOptionsGenerator.REQUEST_OPTIONS_PACKAGE_NAME,
-              RequestBuilderGenerator.REQUEST_OPTIONS_SIMPLE_NAME);
-    }
 
     ParameterizedTypeName requestBuilderOfTranscodeType =
         ParameterizedTypeName.get(
@@ -180,8 +162,9 @@ final class RequestBuilderGenerator {
         .addSuperinterface(Cloneable.class)
         .addMethods(generateConstructors())
         .addMethod(generateDownloadOnlyRequestMethod())
-        .addMethods(generateInstanceMethodOverridesForRequestOptions())
-//        .addMethods(generateGeneratedRequestOptionsEquivalents(generatedOptions))
+        .addMethods(
+            requestOptionsOverrideGenerator.generateInstanceMethodOverridesForRequestOptions(
+                generatedRequestBuilderOfTranscodeType, EXCLUDED_METHODS_FROM_BASE_REQUEST_OPTIONS))
         .addMethods(generateRequestBuilderOverrides())
         .build();
   }
@@ -203,65 +186,6 @@ final class RequestBuilderGenerator {
         });
   }
 
-
-  private List<MethodSpec> generateInstanceMethodOverridesForRequestOptions() {
-    return
-        FluentIterable.from(
-            processorUtil.findInstanceMethodsReturning(
-                baseRequestOptionsType, baseRequestOptionsType))
-            .filter(new Predicate<ExecutableElement>() {
-              @Override
-              public boolean apply(ExecutableElement input) {
-                return !EXCLUDED_METHODS_FROM_BASE_REQUEST_OPTIONS.contains(
-                    input.getSimpleName().toString());
-              }
-            })
-            .transform(
-        new Function<ExecutableElement, MethodSpec>() {
-          @Override
-          public MethodSpec apply(ExecutableElement input) {
-            return generateRequestOptionOverride(input);
-          }
-        })
-        .toList();
-  }
-
-  private MethodSpec generateRequestOptionOverride(ExecutableElement methodToOverride) {
-    MethodSpec.Builder result = ProcessorUtil.overriding(methodToOverride)
-        .returns(generatedRequestBuilderOfTranscodeType)
-        .addModifiers(Modifier.FINAL);
-    result.addCode(
-        CodeBlock.builder()
-            .add(
-                "return ($T) super.$N(",
-                generatedRequestBuilderOfTranscodeType,
-                methodToOverride.getSimpleName())
-            .add(FluentIterable.from(result.build().parameters)
-                .transform(new Function<ParameterSpec, String>() {
-                  @Override
-                  public String apply(ParameterSpec input) {
-                    return input.name;
-                  }
-                })
-                .join(Joiner.on(", ")))
-            .add(");\n")
-            .build());
-
-    if (methodToOverride.getSimpleName().toString().equals("transforms")) {
-      result
-          .addAnnotation(SafeVarargs.class)
-          .addAnnotation(
-              AnnotationSpec.builder(SuppressWarnings.class)
-                  .addMember("value", "$S", "varargs")
-                  .build());
-    }
-
-    for (AnnotationMirror mirror : methodToOverride.getAnnotationMirrors()) {
-      result.addAnnotation(AnnotationSpec.get(mirror));
-    }
-
-    return result.build();
-  }
 
   /**
    * Generates an override of a particular method in {@code com.bumptech.glide.RequestBuilder} that
@@ -305,152 +229,6 @@ final class RequestBuilderGenerator {
               AnnotationSpec.builder(SuppressWarnings.class)
                   .addMember("value", "$S", "varargs")
                   .build());
-    }
-
-    return builder.build();
-  }
-
-  /**
-   * Generates methods with equivalent names and arguments to methods annotated with
-   * {@link GlideOption} in
-   * {@link com.bumptech.glide.annotation.GlideExtension}s that return our generated
-   * {@code com.bumptech.glide.RequestBuilder} subclass.
-   */
-  private List<MethodSpec> generateGeneratedRequestOptionsEquivalents(
-      @Nullable final TypeSpec generatedOptions) {
-    if (generatedOptions == null) {
-      return Collections.emptyList();
-    }
-    return FluentIterable
-        .from(generatedOptions.methodSpecs)
-        .filter(new Predicate<MethodSpec>() {
-          @Override
-          public boolean apply(MethodSpec input) {
-            return isUsefulGeneratedRequestOption(input);
-          }
-        })
-        .transform(new Function<MethodSpec, MethodSpec>() {
-          @Override
-          public MethodSpec apply(MethodSpec input) {
-            return generateGeneratedRequestOptionEquivalent(input);
-          }
-        })
-        .toList();
-  }
-
-  /**
-   * Returns {@code true} if the given {@link MethodSpec} is a useful method to have in our
-   * {@code com.bumptech.glide.RequestBuilder} subclass.
-   *
-   * <p>Only newly generated methods will be included in the generated
-   * {@code com.bumptech.glide.request.BaseRequestBuilder} subclass, so we only have to filter out
-   * methods that override other methods to avoid duplicates.
-   */
-  private boolean isUsefulGeneratedRequestOption(MethodSpec requestOptionMethod) {
-    return
-        !EXCLUDED_METHODS_FROM_BASE_REQUEST_OPTIONS.contains(requestOptionMethod.name)
-        && requestOptionMethod.hasModifier(Modifier.PUBLIC)
-        && !requestOptionMethod.hasModifier(Modifier.STATIC)
-        && requestOptionMethod.returnType.toString()
-            .equals(requestOptionsClassName.toString());
-  }
-
-   /**
-   * Generates a particular method with  an equivalent name and arguments to the given method
-   * from the generated {@code com.bumptech.glide.request.BaseRequestBuilder} subclass.
-   */
-  private MethodSpec generateGeneratedRequestOptionEquivalent(MethodSpec requestOptionMethod) {
-    CodeBlock callRequestOptionsMethod = CodeBlock.builder()
-        .add(".$N(", requestOptionMethod.name)
-        .add(FluentIterable.from(requestOptionMethod.parameters)
-            .transform(new Function<ParameterSpec, String>() {
-              @Override
-              public String apply(ParameterSpec input) {
-                return input.name;
-              }
-            })
-            .join(Joiner.on(", ")))
-        .add(");\n")
-        .build();
-
-    MethodSpec.Builder result = MethodSpec.methodBuilder(requestOptionMethod.name)
-        .addJavadoc(
-            processorUtil.generateSeeMethodJavadoc(requestOptionsClassName, requestOptionMethod))
-        .addModifiers(Modifier.PUBLIC)
-        .varargs(requestOptionMethod.varargs)
-        .addAnnotations(
-            FluentIterable.from(requestOptionMethod.annotations)
-                .filter(new Predicate<AnnotationSpec>() {
-                  @Override
-                  public boolean apply(AnnotationSpec input) {
-                    return !input.type.equals(TypeName.get(Override.class))
-                        // SafeVarargs can only be applied to final methods. GlideRequest is
-                        // non-final to allow for mocking.
-                        && !input.type.equals(TypeName.get(SafeVarargs.class))
-                        // We need to combine warnings below.
-                        && !input.type.equals(TypeName.get(SuppressWarnings.class));
-                  }
-                })
-                .toList()
-        )
-        .addTypeVariables(requestOptionMethod.typeVariables)
-        .addParameters(requestOptionMethod.parameters)
-        .returns(generatedRequestBuilderOfTranscodeType)
-        .beginControlFlow(
-            "if (getMutableOptions() instanceof $T)", requestOptionsClassName)
-        .addCode("this.requestOptions = (($T) getMutableOptions())",
-            requestOptionsClassName)
-        .addCode(callRequestOptionsMethod)
-        .nextControlFlow("else")
-        .addCode(CodeBlock.of("this.requestOptions = new $T().apply(this.requestOptions)",
-            requestOptionsClassName))
-        .addCode(callRequestOptionsMethod)
-        .endControlFlow()
-        .addStatement("return this");
-
-    AnnotationSpec suppressWarnings = buildSuppressWarnings(requestOptionMethod);
-    if (suppressWarnings != null) {
-      result.addAnnotation(suppressWarnings);
-    }
-    return result.build();
-  }
-
-  @Nullable
-  private AnnotationSpec buildSuppressWarnings(MethodSpec requestOptionMethod) {
-    Set<String> suppressions = new HashSet<>();
-    if (requestOptionMethod.annotations.contains(
-        AnnotationSpec.builder(SuppressWarnings.class).build())) {
-      for (AnnotationSpec annotation : requestOptionMethod.annotations) {
-        if (annotation.type.equals(TypeName.get(SuppressWarnings.class))) {
-          List<CodeBlock> codeBlocks = annotation.members.get("value");
-          suppressions.addAll(FluentIterable.from(codeBlocks).transform(
-              new Function<CodeBlock, String>() {
-                @Override
-                public String apply(CodeBlock input) {
-                  return input.toString();
-                }
-              }).toSet());
-        }
-      }
-    }
-
-    if (requestOptionMethod.annotations.contains(
-        AnnotationSpec.builder(SafeVarargs.class).build())) {
-      suppressions.add("unchecked");
-      suppressions.add("varargs");
-    }
-
-    if (suppressions.isEmpty()) {
-      return null;
-    }
-    // Enforce ordering across compilers (Internal and External compilers end up disagreeing on the
-    // order produced by the Set additions above.)
-    ArrayList<String> suppressionsList = new ArrayList<>(suppressions);
-    Collections.sort(suppressionsList);
-
-    AnnotationSpec.Builder builder = AnnotationSpec.builder(SuppressWarnings.class);
-    for (String suppression : suppressionsList) {
-      builder.addMember("value", "$S", suppression);
     }
 
     return builder.build();
